feat(api-gateway): added file for swagger opts
